Switch to float duration and credits_used

+++ src/smooth/models/__init__.py
@@ -252,8 +252,8 @@ class Extension(BaseModel):
 class BaseActionResponse(BaseModel):
   """Base response model for actions."""
 
-  credits_used: int = Field(default=0, description="The amount of credits used to perform the action.")
-  duration: int = Field(default=0, description="The duration taken to perform the action.")
+  credits_used: float = Field(default=0.0, description="The amount of credits used to perform the action.")
+  duration: float = Field(default=0.0, description="The duration taken to perform the action.")
 
 
 class ActionGotoResponse(BaseActionResponse):
